% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Title The Brute Force Knapsnack Solver}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{The input to function containing the number of items}

\item{W}{The weights of the items}
}
\description{
Title The Brute Force Knapsnack Solver
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <- data.frame(w=sample(1:4000, size = n, replace = TRUE), v=runif(n = n, 0, 10000))
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)

}
