% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Title Greedy Algorithm Knapsack}
\usage{
greedy_knapsack(x, W, fast = NULL)
}
\arguments{
\item{x}{The input to function containing the number of items}

\item{W}{The weights of the items}

\item{fast}{Whether to speed up using Rccp}
}
\description{
Title Greedy Algorithm Knapsack
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <- data.frame(w=sample(1:4000, size = n, replace = TRUE), v=runif(n = n, 0, 10000))
greedy_knapsack(x = knapsack_objects[1:8,], W = 3500)

}
