% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Title Dynamic Program Knapsack}
\source{
http://www.es.ele.tue.nl/education/5MC10/Solutions/knapsack.pdf
}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{The input to function containing the number of items}

\item{W}{The weights of the items}
}
\description{
Title Dynamic Program Knapsack
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <- data.frame(w=sample(1:4000, size = n, replace = TRUE), v=runif(n = n, 0, 10000))
knapsack_dynamic(x = knapsack_objects[1:8,], W = 3500)

}
