% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{A dataset containing the knapsnack data is present here}
\format{A data frame with 2000 rows and 2 columns.
\describe{
  \item{w}{weight of the item}
  \item{v}{value of the item}
  ...
}}
\usage{
knapsack_objects
}
\description{
A dataset containing the weight and value of each item
}
\keyword{datasets}
